html/semantics/embedded-content-0/media-elements/canPlayType_method_exists.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.canPlayType() - existence of method</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-navigator-canplaytype">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  !!document.getElementById("a").canPlayType,
  "audioElement.canPlayType method exists");
}, "audio.canPlayType() - existence of method");

test(function() {
 assert_true(
  "canPlayType" in document.getElementById("v"),
  "videoElement.canPlayType method exists");
}, "video.canPlayType() - existence of method");
  </script>
 </body>
</html>
